requirements.txt
# fastapi core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# authentication & security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
boto3==1.34.0

# configuration & environment
python-decouple==3.8
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# database
supabase==2.0.3

# scheduler
APScheduler==3.10.4

# http client
httpx>=0.24.0,<0.25.0
requests==2.31.0

# validation
email-validator==2.1.0

# development & testing
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0